feat(web): Display preliminary instead of estimated when using TSA
