+++ web/src/locales/en.json
@@ -136,6 +136,7 @@
     },
     "estimated_time_slicer_average": {
       "title": "Data is preliminary",
+      "pill": "Preliminary",
       "body": "The data provider is not reporting data for this hour. The displayed values are estimated based on historical measurements and will be updated once measured data is available. Note that estimated values may not account for real-time weather events."
     },
     "estimated_mode_breakdown": {
